refactor(sitesearch): extract shared Retrieve core (text+vector+RRF+rerank)
